% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_wkt.R
\name{poly_wkt}
\alias{poly_wkt}
\title{Make a wkt string from polygon that rgbif will accept}
\usage{
poly_wkt(data, names = NULL)
}
\arguments{
\item{data}{a polygon \code{sf} or \code{sfc} object}
}
\value{
a single polygon wkt string with winding order reversed
}
\description{
gbif takes wkt strings for polygons that are wound anti-clockwise,
while sf polygons are wound clockwise. This function takes an sf polygon
object and converts it to a wkt string with reversed winding order.
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
poly_wkt(nc)

}
